smartspend/backend/requirements.txt
Flask==2.3.2
pymongo==4.7.1
dnspython==2.4.2
python-dotenv==1.0.0
flask-cors==4.0.0
Flask-JWT-Extended==4.4.4
